Some more tidying up (maybe done?)
